Refactor trace to generic compile-time mechanism, clean up casts
- Replace FILE*-based trace with callback-based ms0515_trace_t context
- Add MS0515_TRACE compile-time flag (conan option trace=False, cmake)
- BOARD_TRACE/TRACE macros expand to nothing in release builds
- Remove all debug trace call sites from board.c, cpu.c, cpu_ops.c, ramdisk.c
- Remove trace_io, trace_mem_write, board_trace_dump and related helpers
- Remove --trace CLI argument and trace setup from frontend
- Fix unnecessary int→unsigned casts by declaring variables as proper types
  (floppy.c target, timer.c ch_num/rw, memory.c bank/phys_bank, ramdisk.c bit)
- Use compound assignment (+=/-=) instead of cast in cpu_ops.c byte addressing
- Add zero-warnings rule to CLAUDE.md

+++ emu/core/src/floppy.c
@@ -177,12 +177,12 @@ static void cmd_seek(ms0515_floppy_t *fdc)
     fdc_drive_t *drv = current_drive(fdc);
 
     /* The data register holds the desired track number */
-    int target = fdc->data_reg;
+    uint8_t target = fdc->data_reg;
     if (target >= FDC_TRACKS)
         target = FDC_TRACKS - 1;
 
     drv->track     = target;
-    fdc->track_reg = (uint8_t)target;
+    fdc->track_reg = target;
 
     update_type1_status(fdc);
     finish_command(fdc);
